:recycle: (refactor: step별 style 파일 생성 후 분리)
