+++ src/pages/CreatePostPage/components/FirstStep/style.ts
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+import { Text } from '@refactors/atoms';
+
+const MapHeaderText = styled(Text)`
+  position: absolute;
+  top: 84px;
+  left: 23px;
+  line-height: 1.2;
+  z-index: 1;
+`;
+
+export { MapHeaderText };
